asdf

+++ service/src/main/java/com/comtrade/service/gameservice/GameServiceImpl.java
@@ -51,7 +51,7 @@ public class GameServiceImpl implements GameService {
         if (game.getSlagalicaGame()!=null){
             response.setNumOfPointsSlagalica(game.getSlagalicaGame().getNumOfPoints());
         }if(game.getSpojniceGame() != null){
-            response.setNumOfPointsSpojnice((int) game.getSpojniceGame().getPoints());
+            response.setNumOfPointsSpojnice(game.getSpojniceGame().getPoints());
         }
         if(game.getSkockoGame() != null) {
             response.setNumOfPointsSkocko(game.getSkockoGame().getNumOfPoints());
